requirements-test.txt
fastapi==0.63.0
uvicorn==0.11.5
requests==2.23.0
coverage==5.1
boto3==1.17.67
snowflake-connector-python==2.4.3
mysql-connector==2.2.9
psycopg2-binary==2.8.6
pandas==1.2.4
numpy==1.20.2
moto==2.0.6
pystan==2.19.1.1
prophet==1.0.1
mlflow==1.16.0
plotly==4.14.3